requirements.txt
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
beautifulsoup4 @ file:///tmp/build/80754af9/beautifulsoup4_1601924105527/work
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///opt/concourse/worker/volumes/live/ca3c0fea-9d0f-4489-6889-89b42078ed37/volume/cffi_1613246931142/work
chardet @ file:///opt/concourse/worker/volumes/live/9efbf151-b45b-463d-6340-a5c399bf00b7/volume/chardet_1607706825988/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
cryptography @ file:///opt/concourse/worker/volumes/live/cdf8ff17-b0bf-4081-524c-b5a0afe929ba/volume/cryptography_1616769280208/work
distlib @ file:///home/conda/feedstock_root/build_artifacts/distlib_1622292800270/work
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1589994591731/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///opt/concourse/worker/volumes/live/4e25a0be-45cc-4b73-4314-4604f00e30a4/volume/importlib-metadata_1617877365451/work
pipenv @ file:///home/conda/feedstock_root/build_artifacts/pipenv_1605420760703/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///opt/concourse/worker/volumes/live/ef943889-94fc-4539-798d-461c60b77804/volume/pysocks_1605305801690/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv_1619733560351/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
selenium @ file:///Users/runner/miniforge3/conda-bld/selenium_1610664501146/work
six @ file:///opt/concourse/worker/volumes/live/f983ba11-c9fe-4dff-7ce7-d89b95b09771/volume/six_1605205318156/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1616183228191/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
virtualenv @ file:///Users/runner/miniforge3/conda-bld/virtualenv_1621891696390/work
virtualenv-clone==0.5.4
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
